feat(bex): break out the weights normalization logic into a hub/b-sdk use hook
